pptx_gen/requirements.txt
fastapi
uvicorn[standard]
python-pptx
requests
python-multipart
